requirements.txt
argcomplete==1.10.0
backcall==0.1.0
beautifulsoup4==4.8.0
blis==0.4.1
certifi==2019.9.11
chardet==3.0.4
cymem==2.0.2
decorator==4.4.0
docx2txt==0.8
EbookLib==0.17.1
en-core-web-sm==2.2.0
extract-msg==0.23.1
filetype==1.0.5
idna==2.8
IMAPClient==2.1.0
ipython==7.8.0
ipython-genutils==0.2.0
jedi==0.15.1
lxml==4.4.1
murmurhash==1.0.2
nltk==3.4.5
numpy==1.17.3
olefile==0.46
parso==0.5.1
pdfminer.six==20181108
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.0
pkg-resources==0.0.0
plac==0.9.6
preshed==3.0.2
prompt-toolkit==2.0.10
ptyprocess==0.6.0
pycryptodome==3.9.0
Pygments==2.4.2
python-pptx==0.6.18
pytz==2019.3
regex==2019.8.19
requests==2.22.0
six==1.12.0
sortedcontainers==2.1.0
soupsieve==1.9.4
spacy==2.2.1
spacy-lookups-data==0.1.0
spacy-readability==1.4.1
SpeechRecognition==3.8.1
srsly==0.1.0
syllapy==0.7.0
syntok==1.2.1
textract==1.6.3
thinc==7.1.1
tqdm==4.36.1
traitlets==4.3.3
tzlocal==1.5.1
ujson==1.35
urllib3==1.25.6
wasabi==0.2.2
wcwidth==0.1.7
xlrd==1.2.0
XlsxWriter==1.2.2
